requirements.txt
streamlit==1.28.1
spacy==3.7.2
nltk==3.8.1
anthropic==0.7.8
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
langdetect==1.0.9